[BE] 일부 코드 수정

See merge request s11-mobility-smarthome-sub1/S11P21B209!60

+++ backend/src/main/java/com/safe/safetycar/streaming/Image/ImageManager.java
@@ -13,11 +13,11 @@ public class ImageManager {
     private static ArrayList<Image> images = new ArrayList<>();
 
     public void write(ByteArrayInputStream bis, byte cameraId, byte segNum) {
-        images.get((int)cameraId).write(bis, segNum);
+        images.get(cameraId).write(bis, segNum);
     }
 
     public Image read(byte cameraId) {
-        return images.get((int)cameraId);
+        return images.get(cameraId);
     }
 
     /**
@@ -31,7 +31,7 @@ public class ImageManager {
     }
 
     public void clear(byte cameraId) {
-        images.remove((int)cameraId);
+        images.remove(cameraId);
     }
 
     public void clearAll() {
